% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_attendance.R
\docType{data}
\name{conference_attendance}
\alias{conference_attendance}
\title{AI conference attendance data}
\format{A data frame with 468 rows and 4 variables.
\describe{
\item{conference}{Conference name}
\item{size}{Conference size}
\item{year}{Year}
\item{attendance}{Number of people in attendance}
}}
\source{
Conferences provided the data.
}
\usage{
conference_attendance
}
\description{
Conference attendance over time.
}
\keyword{datasets}
